🌱 Add option to change font size
