+++ src/components/settings/SettingsPage.tsx
@@ -2,42 +2,14 @@ import React from 'react';
 import GoBackButton from 'components/layout/GoBackButton';
 import { Header } from 'components/mdxOverrides';
 import copy from 'consts/copy';
-
-const getItem = key => {
-  try {
-    return localStorage.getItem(key);
-  } catch {}
-};
-
-const setItem = (key, value) => {
-  try {
-    return localStorage.setItem(key, value);
-  } catch {}
-};
+import useFontSize from 'utils/useFontSize';
 
 export default function SettingsPage() {
-  const [fontSize, setFontSize] = React.useState(
-    () => getItem('fontSize') ?? 16,
-  );
-
-  React.useEffect(() => {
-    const root = document.getElementsByTagName('html')[0];
-    root.style.setProperty('font-size', `${fontSize}px`);
-  }, [fontSize]);
+  const { fontSize, increaseFontSize, decreaseFontSize } = useFontSize();
 
   const buttonClass =
     'flex items-center content-center text-white bg-green hover:bg-green-600 rounded-full p-2 w-6 h-6 cursor-pointer';
-  const handleFontDecrease = () => {
-    setFontSize(Number(fontSize) - 2);
-    setItem('fontSize', fontSize);
-  };
-  const handleFontIncrease = () => {
-    setFontSize(Number(fontSize) + 2);
-    setItem('fontSize', fontSize);
-  };
-  // TODO read it on page load
-  // TODO Format buttons
-  // TODO Reset settings
+
   return (
     <>
       <div className="relative whitespace-pre-line">
@@ -51,10 +23,10 @@ export default function SettingsPage() {
           <div className="flex items-center content-center bg-gray-50 bg-opacity-60 p-4 rounded-lg">
             <div>Rozmiar tekstu: {fontSize}px</div>
           </div>
-          <button className={buttonClass} onClick={handleFontDecrease}>
+          <button className={buttonClass} onClick={decreaseFontSize}>
             -
           </button>
-          <button className={buttonClass} onClick={handleFontIncrease}>
+          <button className={buttonClass} onClick={increaseFontSize}>
             +
           </button>
         </div>
